modify

+++ lightning_attn_cute/test.py
@@ -302,9 +302,9 @@ def test_kv_match(k, v, myflash):
 
 
 def test_kv_match_f16(k, v, myflash):
-    torch_kv_output = torch_compute_kv_f16(k, v).to(torch.float32)
+    torch_kv_output = torch_compute_kv_f16(k, v)
     # cute_kv_output = myflash.cute_compute_kv_all_f16(k, v)
-    cute_kv_output = myflash.cute_compute_kv(k, v)
+    cute_kv_output = myflash.cute_compute_kv(k, v).half()
 
     print(f"torch_kv_output dtype: {torch_kv_output.dtype}, cute_kv_output dtype: {cute_kv_output.dtype}")
